==> code/human/notebooks/02_RDS_to_H5AD.ipynb <==
library(reticulate)
library(sceasy)
library(Seurat)


getwd()


seurat_object <- readRDS('/ictstr01/groups/ml01/workspace/shrey.parikh/PDAC/raw_data/Zenodo/Data/pk_all.rds')

seurat_object

# Access the counts matrix from the active assay
raw_counts <- GetAssayData(object = seurat_object, assay = "RNA", layer = "data")

# 1. Check if all counts are integers
are_integers <- all(raw_counts == as.integer(raw_counts))
print(paste("Are raw counts integers?", are_integers))

# # 2. Calculate basic statistics
# # Mean raw counts
# mean_raw_counts <- mean(raw_counts)
# print(paste("Mean raw counts:", mean_raw_counts))

# Range of raw counts
range_raw_counts <- range(raw_counts)
print(paste("Range of raw counts:", range_raw_counts[1], "to", range_raw_counts[2]))

# # Percentage of zero counts
# percentage_zeros <- mean(raw_counts == 0) * 100
# print(paste("Percentage of zero counts:", round(percentage_zeros, 2), "%"))

# Optional: Create a separator for clarity
cat(rep("-", 50), "\n")

#TODO: remove commented code if not necessary

sceasy::convertFormat(seurat_object, from="seurat", to="anndata",
                       outFile='pk_all.h5ad')
