==> code/human/notebooks/04_DataStructure_Plots.ipynb <==
import scanpy as sc
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

adata = sc.read_h5ad("/home/aih/shrey.parikh/PDAC/PDAC/concatenated_datasets/PDAC_concat_hvg_batch_key_datasets_hvg.h5ad")

adata.obs['Condition'] = np.where(adata.obs.Dataset == 'Regev', 'snRNA', 'scRNA')

cell_counts = adata.obs.groupby(['Condition', 'Dataset']).size().reset_index(name='Count')


plt.figure(figsize=(12, 8))  # Increased figure size
barplot = sns.barplot(x='Dataset', y='Count', hue='Condition', data=cell_counts, palette='pastel')

plt.title('Number of Cells in Each Dataset by Condition')
plt.xlabel('Dataset')
plt.ylabel('Number of Cells')
plt.xticks(rotation=45, ha="right")  # Ha aligns the rotation of the x-tick labels
plt.legend(title='Condition')

# Adding annotations on top of each bar, improved
for p in barplot.patches:
    if p.get_height() > 0:  # Skip zero values
        _x = p.get_x() + p.get_width() / 2
        _y = p.get_height()
        value = '{:.0f}'.format(p.get_height())
        barplot.text(_x, _y, value, ha="center", va="bottom", 
                     rotation=0, color='black', 
                     bbox=dict(facecolor='white', edgecolor='white', alpha=0.6))
plt.grid()
plt.tight_layout()
plt.show()



datasets = {
    "Ding": {"Patients": 31, "Samples": 83, "Cells": 261497, "Cell Types": 32},
    "Regev": {"Patients": 48, "Samples": 43, "Cells": 224988, "Cell Types": 33},
    "Simeone": {"Patients": 27, "Samples": 27, "Cells": 139446, "Cell Types": 10},
    "Lee": {"Patients": 9, "Samples": 9, "Cells": 31720, "Cell Types": 8},
    "Steele": {"Patients": 19, "Samples": 16, "Cells": 76798, "Cell Types": "Unknown"},
    "Peng": {"Patients": 35, "Samples": 35, "Cells": 57530, "Cell Types": 10},
    "Zenodo": {"Patients": "Unknown", "Samples": "Unknown", "Cells": 59569, "Cell Types": "Unknown"},
    "Zhang": {"Patients": 4, "Samples": 8, "Cells": 57702, "Cell Types": 12}
}

# Creating a DataFrame
df = pd.DataFrame(datasets).T.reset_index()
df.rename(columns={"index": "Dataset"}, inplace=True)

# Plotting
fig, ax = plt.subplots(figsize=(14, 8))

# Bar plot for number of cells
ax.bar(df["Dataset"], df["Cells"], color='skyblue', label='Number of Cells')

# Adding annotations for cell type counts
for i, row in df.iterrows():
    ax.text(i, row["Cells"], f"{row['Cell Types']} cell types", ha='center', va='bottom', rotation=0, color='black')

# Formatting the plot
ax.set_ylabel('Number of Cells')
ax.set_title('Number of Cells and Cell Types per Dataset')
ax.set_xticklabels(df["Dataset"], rotation=45, ha='right')
plt.grid()
plt.legend()
plt.show()


zenodo =sc.read_h5ad('/home/aih/shrey.parikh/PDAC/PDAC/processed_datasets/All_genes/zenodo_PDAC_new.h5ad')

patient_ids_per_dataset = zenodo.obs.groupby('Dataset')['Patient'].apply(lambda x: x.unique()).reset_index(name='Patient IDs')
patient_ids_per_dataset

zenodo[zenodo.obs.Dataset == 'Schlesinger'].obs.columns

zenodo[zenodo.obs.Dataset == 'Schlesinger'].obs['Patient']

patient_ids_per_dataset['Patient IDs'].apply(lambda x: len(x))

zenodo

for dataset in adata.obs.Dataset.unique():
    temp_adata = adata[adata.obs.Dataset == dataset]
    print(f'{dataset} contains {temp_adata.shape[0]} cells and {temp_adata.shape[1]} genes.')

caronni = sc.read_h5ad('/home/aih/shrey.parikh/PDAC/PDAC/processed_datasets/All_genes/Caronni_concat.h5ad')


datasets = {
    "Ding": {"Patients": 31, "Samples": 83, "Cells": 261497, "Cell Types": 32},
    "Regev": {"Patients": 48, "Samples": 43, "Cells": 224988, "Cell Types": 33},
    "Simeone": {"Patients": 27, "Samples": 27, "Cells": 139446, "Cell Types": 10},
    "Lee": {"Patients": 9, "Samples": 9, "Cells": 31720, "Cell Types": 8},
    "Steele": {"Patients": 19, "Samples": 16, "Cells": 76798, "Cell Types": "Unknown"},
    "Peng": {"Patients": 35, "Samples": 35, "Cells": 57530, "Cell Types": 10},
    "Zenodo_OUGs": {"Patients": 4, "Samples": "Unknown", "Cells": 19797, "Cell Types": 10},
    "Zhang": {"Patients": 4, "Samples": 8, "Cells": 57702, "Cell Types": 12},
    "Caronni": {"Patients": 4, "Samples": 4, "Cells": 57702, "Cell Types": 12},
    
}

# Creating a DataFrame
df = pd.DataFrame(datasets).T.reset_index()
df.rename(columns={"index": "Dataset"}, inplace=True)

# Plotting
fig, ax = plt.subplots(figsize=(14, 8))

# Bar plot for number of cells
ax.bar(df["Dataset"], df["Samples"], color='skyblue', label='Number of Donors')

# # Adding annotations for cell type counts
# for i, row in df.iterrows():
#     ax.text(i, row["Cells"], f"{row['Cell Types']} cell types", ha='center', va='bottom', rotation=45, color='black')

#TODO: remove commented code if not necessary

# Formatting the plot
ax.set_ylabel('Number of Cells')
ax.set_title('Number of Cells and Cell Types per Dataset')
ax.set_xticklabels(df["Dataset"], rotation=45, ha='right')

plt.legend()
plt.show()


Ding contains 211138 cells and 4260 genes.
Lee contains 8897 cells and 4260 genes.
Regev contains 192044 cells and 4260 genes.
Simeone contains 213711 cells and 4260 genes.
Steele contains 37218 cells and 4260 genes.
Peng contains 41964 cells and 4260 genes.
Caronni contains 75280 cells and 4260 genes.
Zhang contains 37934 cells and 4260 genes.
Moncada contains 3625 cells and 4260 genes.
Schlesinger contains 4874 cells and 4260 genes.
Zenodo_OUGS contains 11298 cells and 4260 genes.
#TODO: remove?

simeone = sc.read_h5ad('/home/aih/shrey.parikh/PDAC/PDAC/processed_datasets/All_genes/Simeone_concat.h5ad')


cells_have = {
    'Caronni': 75280,
    'Ding': 211138,
    'Lee': 8897,
    'Moncada': 3625,
    'Peng': 41964,
    'Regev': 192044,
    'Schlesinger': 4874,
    'Simeone': 213711,
    'Steele': 37218,
    'Zenodo_OUGs': 11298,
    'Zhang': 37934
}

cells_quoted = {
    'Caronni': 59569,
    'Ding': 232764,
    'Lee': 31720,
    'Moncada': 3625,
    'Peng': 41986,
    'Regev': 224988,
    'Schlesinger': 4874,
    'Simeone': 139446,
    'Steele': 46244,
    'Zenodo_OUGs': 19797,
    'Zhang': 57702
}

# Sort keys to ensure the bars match up
datasets = sorted(cells_have.keys())

# Collect the corresponding values, ensuring the keys match between the two dictionaries
have_values = [cells_have[dataset] for dataset in datasets]
quoted_values = [cells_quoted[dataset] for dataset in datasets]

# Plotting
fig, ax = plt.subplots(figsize=(14, 8))

# Width of a bar
bar_width = 0.4

# Positions of the bars
r1 = np.arange(len(have_values))
r2 = [x + bar_width for x in r1]

# Create bars for the cells we have
bars_have= ax.bar(r1, have_values, color='skyblue', width=bar_width, label='Cells Obtained')

# Create bars for the cells quoted in the paper
bars_quoted = ax.bar(r2, quoted_values, color='salmon', width=bar_width, label='Cells Quoted in Paper')

# Function to add value labels on the bars
def add_value_labels(bars):
    for bar in bars:
        height = bar.get_height()
        ax.annotate('{}'.format(height),
                    xy=(bar.get_x() + bar.get_width() / 2, height),
                    xytext=(0, 3),  # 3 points vertical offset
                    textcoords="offset points",
                    ha='center', va='bottom')

# Call the function for each set of bars
add_value_labels(bars_have)
add_value_labels(bars_quoted)

# Add labels, title, and legend
ax.set_xlabel('Dataset', fontweight='bold')
ax.set_ylabel('Number of Cells', fontweight='bold')
ax.set_title('Comparison of Cell Counts: Obtained vs. Quoted')
ax.set_xticks([r + bar_width/2 for r in range(len(have_values))])
ax.set_xticklabels(datasets, rotation=45, ha='right')
ax.legend()
plt.grid()
# Show the plot
plt.tight_layout()
plt.show()


# Grouping by 'Dataset' and counting unique 'ID' values
patient_counts = adata.obs.groupby('Dataset')['ID'].nunique().reset_index(name='Number of Patients')

# Plotting
plt.figure(figsize=(10, 6))
barplot = plt.bar(patient_counts['Dataset'], patient_counts['Number of Patients'], color='lightblue')

# Add labels and title
plt.xlabel('Dataset')
plt.ylabel('Number of Patients')
plt.title('Number of Unique Patients in Each Dataset')

# Adding the text labels above the bars
for bar in barplot:
    yval = bar.get_height()
    plt.text(bar.get_x() + bar.get_width() / 2, yval + 0.5, yval, ha='center', va='bottom')

# Rotate x-axis labels for better readability
plt.xticks(rotation=45, ha='right')

# Show the plot
plt.grid()
plt.tight_layout()
plt.show()

datasets = {
    "Ding": {"Patients": 33,  "Cells": 261497, "Cell Types": 32},
    "Regev": {"Patients": 43,  "Cells": 224988, "Cell Types": 33},
    "Simeone": {"Patients": 27,  "Cells": 139446, "Cell Types": 10},
    "Lee": {"Patients": 5, "Cells": 31720, "Cell Types": 8},
    "Steele": {"Patients": 16,  "Cells": 76798, "Cell Types": 16},
    "Peng": {"Patients": 24, "Cells": 57530, "Cell Types": 10},
    "Zenodo_OUGs": {"Patients": 4,  "Cells": 19797, "Cell Types": 10},
    "Zhang": {"Patients": 3, "Cells": 57702, "Cell Types": 12},
    "Caronni": {"Patients": 10,  "Cells": 57702, "Cell Types": 12},
    "Moncada": {"Patients": 2,  "Cells": 3625, "Cell Types": 8},
    "Schlesinger": {"Patients": 1, "Cells": 4874, "Cell Types": 10}
    
}


# Data for plotting
cell_types_counts = [datasets[dataset]["Cell Types"] for dataset in datasets]
datasets_labels = list(datasets.keys())

# Create a pie chart
fig, ax = plt.subplots()
ax.pie(cell_types_counts, labels=datasets_labels, autopct=lambda p: '{:.0f}'.format(p * sum(cell_types_counts) / 100),
        startangle=90, counterclock=False)

# Add a legend and title
ax.legend(datasets_labels, title="Datasets", loc="center left", bbox_to_anchor=(1, 0, 0.5, 1))
plt.title('Distribution of Cell Types Across Datasets')

# Show the plot
plt.tight_layout()
plt.show()

adata = sc.read_h5ad("/home/aih/shrey.parikh/PDAC/PDAC/concatenated_datasets/PDAC_concat_hvg_batch_key_donors_hvg_final.h5ad")

sc.pp.normalize_total(adata, target_sum=1e4)
sc.pp.log1p(adata)


adata_qc = adata.copy()

sc.pp.calculate_qc_metrics(adata_qc, inplace=True)

datasets = adata_qc.obs['Dataset'].unique()
dataset_subsets = {ds: adata_qc[adata_qc.obs['Dataset'] == ds] for ds in datasets}

counts = {ds: subset.obs_vector("total_counts") for ds, subset in dataset_subsets.items()}

sc.pl.umap(adata_qc, color='Dataset', frameon=False)

fig, ax = plt.subplots(figsize=(10, 6))

colors = adata_qc.uns['Dataset_colors']

for (ds, count_data), color in zip(counts.items(), colors):
    # Compute histogram data
    hist_data, bin_edges = np.histogram(count_data, bins=30, density=True)
    bin_centers = (bin_edges[:-1] + bin_edges[1:]) / 2

    # Plot as a line
    ax.plot(bin_centers, hist_data, label=ds, color=color)

ax.set_xlabel('Total Counts per Cell')
ax.set_ylabel('Density')
ax.set_title('Normalized Count Distribution by Dataset')
ax.legend(title='Dataset')

plt.show()

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()
scaled_counts = {ds: scaler.fit_transform(count_data.reshape(-1, 1)).flatten() for ds, count_data in counts.items()}

fig, ax = plt.subplots(figsize=(10, 6))

colors = adata_qc.uns['Dataset_colors']

for (ds, count_data), color in zip(scaled_counts.items(), colors):
    # Compute histogram data
    hist_data, bin_edges = np.histogram(count_data, bins=30, density=True)
    bin_centers = (bin_edges[:-1] + bin_edges[1:]) / 2

    # Plot as a line
    ax.plot(bin_centers, hist_data, label=ds, color=color)

ax.set_xlabel('Total Counts per Cell')
ax.set_ylabel('Density')
ax.set_title('Normalized Count Distribution by Dataset')
ax.legend(title='Dataset')

plt.show()
