==> code/human/notebooks/04_Explore.ipynb <==
import os
import scanpy as sc
import seaborn as sb
import numpy as np
import matplotlib.pyplot as plt
import warnings
import anndata as ad
import pandas as pd
sc.settings.n_jobs=190

ding = sc.read_h5ad('Ding/data/Liding_preprocessed.h5ad')

ding.obs['Dataset'] = 'Ding'

ding.obs['sampleid'] = ding.obs['sampleid'].str.replace('HT056P', 'HT056P1')

ding.obs.rename(columns={'sampleid': 'ID'}, inplace=True)

ding.write('Ding/data/Ding_PDAC.h5ad')

not tumour: HT166P1 HT168P1 HT190P1 HT200P1
#TODO: remove?

sampleid_list = [
    'HT056P1', 'HT060P1', 'HT061P1', 'HT064P1', 'HT071P1',
    'HT085P1', 'HT115P1', 'HT121P1', 'HT122P1', 'HT123P1',
    'HT124P1', 'HT125P1', 'HT138P1', 'HT140P1', 'HT166P1',
    'HT168P1', 'HT185P1', 'HT190P1', 'HT191P1', 'HT200P1',
    'HT204P1', 'HT264P1', 'HT270P1', 'HT284P1', 'HT288P1',
    'HT224P1', 'HT231P1', 'HT232P1', 'HT259P1', 'HT306P1',
    'HT242P1', 'HT412P1'
]

[i for i in ding.obs.sampleid.unique() if i not in sampleid_list]

ding = sc.read_h5ad('Ding/data/Ding_PDAC.h5ad')

Lee = sc.read_h5ad('Lee/Lee_concat.h5ad')

Lee_PDAC = Lee[Lee.obs.ID.isin(['P1_filtered_feature_bc_matrix',
 'P2_filtered_feature_bc_matrix',
 'P3_filtered_feature_bc_matrix',
 'P4_filtered_feature_bc_matrix',
 'P5_filtered_feature_bc_matrix'])]

#written to new folder, change made on 29.01.2024
# Lee_PDAC.write('Lee/Lee_concat_PDAC.h5ad')
Lee_PDAC.write('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/Processed_datasets/All_genes/Lee_concat_PDAC.h5ad')
#TODO: remove commented code if not necessary

Lee_PDAC = sc.read_h5ad('Lee/Lee_concat_PDAC.h5ad')

regev = sc.read_h5ad('Regev_SN/final/GSE202051_totaldata-final-toshare_preprocessed.h5ad')

regev_PDAC = regev

regev_PDAC.obs['ID'] = regev_PDAC.obs.pid
regev_PDAC.obs['Dataset'] = 'Regev'

regev_PDAC.write('Regev_SN/final/Regev_PDAC.h5ad')

regev_PDAC = sc.read_h5ad('Regev_SN/final/Regev_PDAC.h5ad')
#TODO: necessary to write and read?

simeone = sc.read_h5ad('Simeone/Simeone_concat.h5ad')

steele = sc.read_h5ad('Steele/Steele_concat.h5ad')

steele_PDAC = steele[steele.obs.ID.str.contains('PDAC_TISSUE')]

#written to Processed_data folder, change made on 29.01.2024
# steele.write('Steele_concat_PDAC.h5ad')
steele_PDAC.write('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/Processed_datasets/All_genes/Steele_concat_PDAC.h5ad')
#TODO: remove commented code if not necessary

# steele_PDAC = sc.read_h5ad('Steele/Steele_concat_PDAC.h5ad')
#TODO: remove commented code if not necessary

adata_list = [ding, Lee_PDAC, regev_PDAC, simeone, steele_PDAC]

PDAC_initial_int.obs.Dataset.unique()

PDAC_initial_int.write('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/PDAC_initial_int.h5ad')
#TODO: PDAC_initial_int not defined

full_latent = sc.read_h5ad('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/PDAC_ref+query_latent_scVI_t1.h5ad')

peng_raw = sc.read_h5ad('/lustre/groups/ml01/workspace/shrey.parikh/PDAC/raw_data/Peng/Raw_Data/StdWf1_PRJCA001063_CRC_besca2.raw.h5ad')

zhang = sc.read_h5ad('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/raw_data/Zhang/Zhang_concat.h5ad')

zhang_PDAC = zhang[zhang.obs.ID.str.contains('PDAC')]

zhang_PDAC.write('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/raw_data/Zhang/Zhang_concat_PDAC.h5ad')

zenodo = sc.read_h5ad('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/raw_data/Zenodo/Data/pk_all.h5ad')

CA001063 = Peng (used)
GSE111672 = Moncada (new)
GSE154778 = Lee (used)
GSE155698 = Steele (used)
GSM4293555 = Schlesinger (new)
OUGS = Zenodo paper (new)

zenodo_list = ['GSE111672', 'GSM4293555', 'OUGS']

zenodo_PDAC_new = zenodo[zenodo.obs.Project.isin(zenodo_list)]

zenodo_PDAC_new.write('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/raw_data/Zenodo/Data/zenodo_PDAC_new.h5ad')

zenodo_PDAC_new.var.rename(columns={'name': '1'}, inplace=True)

zenodo_PDAC_new.obs['Dataset'] = zenodo_PDAC_new.obs.Project.map({'GSE111672': 'Moncada', 'GSM4293555': 'Schlesinger', 'OUGS' : 'Zenodo_OUGS'}).astype('category')

zenodo_PDAC_new.obs['ID'] = zenodo_PDAC_new.obs.Patient

zenodo_PDAC_new[:,zenodo_PDAC_new.var.index.str.contains('Jan|Feb|Mar|April|May|Jun|Jul|Auh|Sep|Oct|Nov|Dec')].var

months_pattern = 'Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec'
zenodo_PDAC_new_filtered = zenodo_PDAC_new[:, ~zenodo_PDAC_new.var.index.str.contains(months_pattern)]

zenodo_PDAC_new_filtered.write('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/raw_data/Zenodo/Data/zenodo_PDAC_new.h5ad')

zenodo_PDAC_new_filtered = sc.read_h5ad('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/raw_data/Zenodo/Data/zenodo_PDAC_new.h5ad')

peng = sc.read_h5ad('Peng/Processed_Data/StdWf1_PRJCA001063_CRC_besca2.annotated.h5ad')

peng_raw = sc.read_h5ad('Peng/Raw_Data/StdWf1_PRJCA001063_CRC_besca2.raw.h5ad')

peng.obs['Dataset'] = 'Peng'
peng.obs['ID'] = peng.obs.Patient

peng_PDAC = peng[peng.obs.ID.str.contains('T')]

peng_PDAC.write('Peng/Processed_Data/peng_PDAC.h5ad')

peng_raw.obs['Dataset'] = 'Peng'
peng_raw.obs['ID'] = peng_raw.obs.Patient

peng_PDAC = peng_raw[peng_raw.obs.index.isin(peng.obs.index.tolist())]

peng_PDAC = peng_PDAC[peng_PDAC.obs.ID.str.contains('T')]

# peng_PDAC.write('Peng/Processed_Data/peng_PDAC.h5ad')
peng_PDAC.write('/dss/dssfs02/lwp-dss-0001/pn36po/pn36po-dss-0001/di35nod/PDAC_data/Processed_datasets/All_genes/peng_PDAC.h5ad')

sc.pp.neighbors(regev)

sc.tl.umap(regev)

sc.pl.umap(regev, color=['Level 2 Annotation', 'pid'],wspace=0.75)

#TODO: are the next 3 cells necessary?
data = sc.read_h5ad('/lustre/groups/ml01/workspace/shrey.parikh/PDAC/raw_data/Zenodo/Data/pk_all.h5ad')

data[data.obs.Patient2.str.startswith('OUGS')].obs.groupby(['Patient', 'Patient2']).size()

import re
sorted_data = sorted(data, key=lambda x: (int(re.findall(r'\d+', x)[0]), x))

sorted_data
