==> code/human/notebooks/11_Create_MG.ipynb <==
import scanpy as sc
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import warnings
warnings.filterwarnings('ignore')
import anndata as ad

binary_matrix = pd.read_csv('../MOFA/MOFA_15_Factors_6082HVG/binary_matrix_factors.csv', index_col='Unnamed: 0')
adata = sc.read_h5ad('../MOFA/adata_combined_even_more_hvg.h5ad')

genes_mofa_factors = binary_matrix.T.columns.tolist()

pd.DataFrame(genes_mofa_factors).to_csv('MOFA_genes.csv')

len(genes_mofa_factors)

# DE genes across Technologies

sc.tl.rank_genes_groups(adata, groupby='Condition', layer='log_norm')

sc.pl.rank_genes_groups_dotplot(adata, groupby='Condition', values_to_plot='logfoldchanges', n_genes=100)

df = pd.DataFrame(adata.uns['rank_genes_groups']['names'])

df_sc = sc.get.rank_genes_groups_df(adata, group='scRNA-seq')
df_sn = sc.get.rank_genes_groups_df(adata, group='snRNA-seq')

de_sc = df_sc[df_sc.logfoldchanges>5].names.tolist()
de_sn = df_sn[df_sn.logfoldchanges>5].names.tolist()

sc_remove = ([i for i in de_sc if i in genes_mofa_factors])
sn_remove = ([i for i in de_sn if i in genes_mofa_factors])

genes_to_remove_techincal_variability= sn_remove + sc_remove

len(genes_to_remove_techincal_variability)

sc.pl.dotplot(adata, groupby='Level_1', var_names=genes_to_remove_techincal_variability,standard_scale='var', layer='log_norm' )

# Markers from Broad

broad_markers = pd.read_csv('broad_markers.csv', index_col='Unnamed: 0')

broad_markers

broad_markers_list = list(set(broad_markers.values.flatten().astype(str)[broad_markers.values.flatten().astype(str) != 'nan']))

pd.DataFrame(broad_markers_list).to_csv('broad_markers_list.csv')

([i for i in broad_markers_list if i not in adata.var_names])

# DE Markers across Cell Types

sc.tl.rank_genes_groups(adata, groupby='Level_1', layer='log_norm')

df = pd.DataFrame(adata.uns['rank_genes_groups']['names']).head(30)
print(len(list(set(df.values.flatten()))))

# Filter out genes that are already present in genes_mofa_factors 
genes_to_check = ([i for i in (list(set(df.values.flatten()))) if i not in genes_mofa_factors])
print(len(genes_to_check))

df.head()

gene_to_group = {}
for gene in genes_to_check:
    test= df.where(df==gene).dropna(how='all').dropna(axis=1, how='all')
    cols = test.columns.tolist()
    gene_to_group[gene] = cols

df['Adipocyte'][df['Adipocyte'].isin(genes_to_check)].tolist()

gene_to_group_T = {}
for cols in df.columns:
    temp = df[cols]
    temp_genes = temp[temp.isin(genes_to_check)].tolist()
    gene_to_group_T[cols] = temp_genes

df = pd.DataFrame.from_dict(gene_to_group_T, orient='index')

de_genes = list(set([val for val in df.values.flatten() if val is not None]))
print(type(de_genes))

df_genes = pd.DataFrame(de_genes, columns=['Gene'])

df_genes.reset_index(inplace=True)

print(df_genes.dtypes)

df_genes.to_csv('de_genes_to_be_added.csv')

from collections import Counter
Counter(pd.DataFrame.from_dict(gene_to_group, orient='index').values.flatten())

# Check Genes

all_genes = (genes_mofa_factors + broad_markers_list + de_genes)

len([i for i in list(set(all_genes)) if i not in adata.var_names])

[i for i in all_genes if i not in adata.var_names]

# Reload

mofa_genes_df = pd.read_csv('MOFA_genes.csv')
broad_markers = pd.read_csv('broad_markers.csv', index_col='Unnamed: 0')
broad_markers_list = list(set(broad_markers.values.flatten().astype(str)[broad_markers.values.flatten().astype(str) != 'nan']))
de_genes_df = pd.read_pickle('de_genes_to_be_added.csv')
mofa_genes = mofa_genes_df.values.flatten().tolist()
de_genes = de_genes_df.values.flatten().tolist()

all_genes = (genes_mofa_factors + broad_markers_list + de_genes)

# Xenium Panel

xenium = pd.read_csv('pdac_xenium_panel.csv')

xenium_genes = list(set(xenium.Gene.tolist()))

len(xenium_genes)

# check how many genes not in all_genes
len([i for i in xenium_genes if i not in all_genes])

valid_genes = [i for i in xenium_genes if i in adata.var_names]

genes_to_plot = [i for i in valid_genes if i not in all_genes]

# Number of xenium genes present in 6082 HVG and in the selected all_genes list
len(genes_to_plot)

sc.pl.dotplot(adata, groupby='Level_1', var_names=genes_to_plot, layer='log_norm', standard_scale='var')

# Create final list

all_genes = list(set(mofa_genes + broad_markers_list + de_genes + xenium_genes))

len(all_genes)

all_genes = list(set(mofa_genes + broad_markers_list + de_genes + xenium_genes))

all_genes = [i for i in all_genes if i in adata.var_names]
