==> code/human/notebooks/14_Biological_Metadata.ipynb <==
import pandas as pd
import numpy as np

pwd

pd.set_option('display.max_columns', None)

df_id = pd.read_csv('ID_Harmonised.csv')

# Ding

ding = pd.read_excel('metadata_publications/Ding_metadata.xlsx', sheet_name='Table2_Metadata')

df_id

ding.tail()

ding.columns

ding_dedup = ding[ding.Case.isin(df_id.ID_harmonised.tolist())].drop_duplicates(subset='Case', keep='first')


interesing_cols = ['Case', 'KRAS_mutation', 'KRAS_type', 'TP53_type', 'SMAD4_type',
       'CDKN2A_type', 'Germline_ATM', 'Germline_BRCA2', 'Germline_FANCC', 'Moffitt_Subtype', 'Bailey_Subtype', 'Collisson_Subtype']

df_id_merged = df_id.merge(ding_dedup[interesing_cols], left_on='ID_harmonised', right_on='Case', how='left')

df_id_merged.drop('Dataset_Barcode', axis=1, inplace=True)

df_id_merged.drop_duplicates()

ding_2 = pd.read_excel('metadata_publications/Ding_metadata.xlsx', sheet_name='Table1_Clinical_data')

ding_2_dedup = ding_2[ding_2.Patient_ID.isin(df_id.ID_harmonised.tolist())].drop_duplicates(subset='Patient_ID', keep='first')

ding_2_dedup.head()

ding_2_dedup.columns

# Function to extract the T stage and N stage
def extract_tnm(row):
    # Extract T stage from Pathology_notes
    t_stage = row['Pathology_notes'].split(',')[0]  # Get the first part before the comma
    # Extract N stage from Lymph_node_met
    n_stage = row['Lymph_node_met'].split(',')[1].strip()  # Get the part after the comma
    return f"{t_stage} {n_stage}"

# Create the new TNM classification column
ding_2_dedup['TNM_classification'] = ding_2_dedup.apply(extract_tnm, axis=1)

interesing_cols = ['Patient_ID', 'Time_from_diagnosis_to_surgery_(weeks)', 'Cycles', 'Location', 'Grade', 'Size_(cm)', 'Stage','Diabetes', 
                   'Pancreatitis', 'Previous_cancer_history',  'Other_co-morbidities', 'Smoker', 'Alcohol', 'Pathology_notes', 'Lymph_node_met',
                   'Survival_post_surgery_(weeks)', 'Vital_status', 'TNM_classification']

ding_2_dedup.head(2)

df_id_merged = df_id_merged.merge(ding_2_dedup[interesing_cols], left_on='ID_harmonised', right_on='Patient_ID', how='left')

df_id_merged.head()

df_id_merged.shape

# Hwang

hwang = pd.read_excel('metadata_publications/Hwang_metadata.xlsx', sheet_name='Supplementary Table 1', header=1)

hwang.head(5)

hwang['Vital_status'] = hwang['Status Last FUP'].map({
    'DWD': 'Deceased - Disease',
    'DWOD': 'Deceased - Other',
    'MET': 'Alive - Metastatic',
    'NED': 'Alive - Disease-free'
})

hwang[['TNM_classification', 'Grade']] = hwang['Stage/Grade'].str.split('/', expand=True)

hwang.rename(columns={'PFS (d)': 'Progression-Free Survival', 'OS (d)': 'Overall Survival'}, inplace=True)

hwang.head()

df_id_merged[df_id_merged.Dataset.str.contains('Hwa')]

df_id_merged[df_id_merged.Dataset.str.contains('Hwa')].shape

hwang_clean = hwang[hwang['ID'].notna()].copy()

hwang_clean['ID'] = (
    hwang_clean['ID']
    .astype(str)                   # ensure all are strings
    .str.strip()                   # remove leading/trailing spaces
    .str.replace('†', '', regex=False)  # remove dagger symbol
)

hwang_clean = hwang_clean[hwang_clean['ID'].str.startswith('PDAC_')]

hwang_clean.columns

# df_id_merged['Overall Survival'] 

df_id_merged['Overall Survival'] = df_id_merged['Survival_post_surgery_(weeks)'] * 7

# interesing_cols = ['ID', 'Grade', 'TNM_classification' 'Vital_status', 'Overall Survival']

len(set(hwang_clean.ID) & set(df_id_merged.ID_harmonised))

hwang_map = hwang_clean.set_index('ID')

interesing_cols = ['Grade','TNM_classification', 'Vital_status', 'Overall Survival']

for col in interesing_cols:
    if col in df_id_merged.columns:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(hwang_map[col]).fillna(df_id_merged[col])
    else:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(hwang_map[col])    

df_id_merged.head(3)

df_id_merged.groupby(['Dataset', 'Grade']).size().unstack().T

df_id_merged.groupby(['Dataset', 'TNM_classification']).size().unstack().T

# Simeone

simeone = pd.read_excel('metadata_publications/Simeone_metadata.xlsx', sheet_name='Genomic Mutations')

simeone #.head(3)

simeone.rename(columns={'KRAS': 'KRAS_mutation', 'TP53': 'TP53_type', 'CDKN2A': 'CDKN2A_type', 'SMAD4': 'SMAD4_type', 'Source': 'Tissue'}, inplace=True)

simeone.columns

df_id_merged[df_id_merged.Dataset.str.contains('Simeo')].head(2)

interesing_cols = ['Procedure', 'KRAS_mutation',
       'TP53_type', 'CDKN2A_type', 'SMAD4_type']

simeone_map = simeone.set_index('Patient')

simeone_map = simeone_map.iloc[:27, :]

simeone_map

simeone_map['Unique_ID'] = simeone_map.index.map(dict(zip(df_id_merged[df_id_merged.Dataset.str.contains('Simeo')].ID_harmonised, df_id_merged[df_id_merged.Dataset.str.contains('Simeo')].Unique_ID)))

simeone_map = simeone_map[~simeone_map.Unique_ID.isna()].set_index('Unique_ID')

for col in interesing_cols:
    if col in df_id_merged.columns:
        df_id_merged[col] = df_id_merged['Unique_ID'].map(simeone_map[col]).fillna(df_id_merged[col])
    else:
        df_id_merged[col] = df_id_merged['Unique_ID'].map(simeone_map[col])    

df_id_merged[df_id_merged.Dataset.str.contains('Simeo')].head(2)

df_id_merged.groupby(['Dataset', 'KRAS_mutation']).size().unstack().T

# Steele

import scanpy as sc

pwd

steele_atlas = sc.read_h5ad('../../Steele_Atlas.h5ad')

steele_atlas.obs.head(2)

steele_atlas.obs[steele_atlas.obs['GSE.SRA..Study.'].str.contains('GSE155698')].head(2)

steele = sc.read_h5ad('/lustre/groups/ml01/workspace/shrey.parikh/PDAC/raw_data/Steele/PDAC_TISSUE_1/filtered_feature_bc_matrix_output.h5ad')

steele.obs

# Steele atlas is useless again

# Check the zenodo 

zenodo = sc.read_h5ad('/lustre/groups/ml01/workspace/shrey.parikh/PDAC/raw_data/Zenodo/Data/pk_all.h5ad')

zenodo.obs[zenodo.obs.Project == 'GSE155698'].isna().sum()

### PDAC
- All samples are treatment naive
- Average age 60.2 (47-72)
- Std Dev BMI 23.5 (4.24)
### Adj_Norm
- Average age 66.4 (45-81)
- Std Dev BMI 25 (5.54)

conditions = [
    (df_id_merged['Dataset'] == 'Steele'),
    (df_id_merged['Dataset'] == 'Steele_Adj_Norm')
]
values = [60.2, 66.4]
df_id_merged['Age'] = np.select(conditions, values, default=np.nan)

conditions = [
    (df_id_merged['Dataset'] == 'Steele'),
    (df_id_merged['Dataset'] == 'Steele_Adj_Norm')
]
values = [23.5, 25]
df_id_merged['BMI'] = np.select(conditions, values, default=np.nan)

df_id_merged['Treatment'] = np.where(df_id_merged.Dataset == 'Steele', 'Treatment Naive', np.nan)

df_id_merged.groupby(['Dataset', 'BMI']).size().unstack()

df_id_merged.groupby(['Dataset', 'Treatment']).size().unstack()

# Lee

 P2-P5 primary samples came from treatment-naïve patients, while P1 was a second FNA from a patient who had been treated with gemcitabine/paclitaxel. 
For metastatic samples, the most recent therapies prior to sample acquisition were 5-fluorouracil/liposmal irinotecan, evofosfamide/ipilimumab and 
capecitabine for VM, LuM, and PM, respectively. 

df_id_merged[df_id_merged.Dataset == 'Lee']

### Added already by Daniele

# Peng

%pip install xlrd

peng = pd.read_excel('metadata_publications/Peng_metadata.xlsx', header=1)

peng.columns

peng.head(2)

df_id_merged[df_id_merged.Dataset == 'Peng'].head(2)

peng.rename(columns={'pathologic diagnoses': 'Pathology_notes', 'Staging': 'Stage', 'TNM Classification': 'TNM_classification', 'Age(y/o)': 'Age'}, inplace=True)

peng.columns

interesing_cols = ['Pathology_notes', 'Age', 'Procedure', 
       'TNM_classification',  'Stage', 'perineural invasion',
       'vascular invasion', 'peripancreatic infiltarion']

peng[['Random', 'Unique_ID']] = peng.Number.str.split('-', expand=True)

df_id_merged.ID_harmonised = np.where(df_id_merged.Dataset.str.contains('Peng'), 'Peng' + '_' + df_id_merged.ID_harmonised, df_id_merged.ID_harmonised)

# df_id_merged.ID_harmonised = df_id_merged.ID_harmonised.str.replace(r'^Peng_Peng_Peng_', 'Peng_', regex=True)

# df_id_merged.ID_harmonised = df_id_merged.ID_harmonised.str.replace(r'^Peng_Peng_', 'Peng_', regex=True)

df_id_merged[df_id_merged.Dataset.str.contains('Peng')].head(2)

peng.Unique_ID = 'Peng' + '_' + peng.Unique_ID

len(set(peng.Unique_ID))

len(set(peng.Unique_ID) & set(df_id_merged.ID_harmonised))

peng_map = peng.iloc[:40, :]

len(set(peng_map.Unique_ID) & set(df_id_merged.ID_harmonised))

peng_map = peng_map[~peng_map.Unique_ID.isna()].set_index('Unique_ID')

peng_map.head(2)

for col in interesing_cols:
    if col in df_id_merged.columns:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(peng_map[col]).fillna(df_id_merged[col])
    else:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(peng_map[col])    

df_id_merged.groupby(['Dataset', 'TNM_classification']).size().unstack().T

# Caronni

df_id_merged.Dataset.unique()

# columns_to_load = ['ID', 'age', 'sex', 'Site of lesion', 'Chemotherapy schedule', 'Grading', 'TNM']
caronni = pd.read_excel('metadata_publications/Caronni_metadata.xlsx', 
                         sheet_name='Patients clinical data')


df_id_merged.Location

caronni = caronni[~caronni.ID.isna()]

caronni.rename(columns={'Site of lesion':'Location', 'TNM':'TNM_classification', 'Grading': 'Grade'}, inplace=True)

caronni.head()

df_id_merged[df_id_merged.Dataset == 'Caronni']

len(set(caronni.ID) & set(df_id_merged.ID_harmonised))

caronni_map = caronni.set_index('ID')

interesing_cols = ['Location', 'Grade', 'TNM_classification', 'Tissue']

for col in interesing_cols:
    if col in df_id_merged.columns:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(caronni_map[col]).fillna(df_id_merged[col])
    else:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(caronni_map[col])    

df_id_merged.groupby(['Dataset', 'TNM_classification']).size().unstack()

# Linn

linn = pd.read_excel('metadata_publications/Lin_metadata.xlsx')

linn

df_id_merged.ID_harmonised = np.where(df_id_merged.Dataset.str.contains('Lin'), 'Lin' + '_' + df_id_merged.ID_harmonised, df_id_merged.ID_harmonised)

df_id_merged[df_id_merged.Dataset.str.contains('Lin')]

linn['Patient ID'] = 'Lin' + '_' + linn['Patient ID']

len(linn['Patient ID'])

len(set(linn['Patient ID']) & set(df_id_merged.ID_harmonised))

interesing_cols = ['Stage', 'Grade']

linn_map = linn.set_index('Patient ID')

for col in interesing_cols:
    if col in df_id_merged.columns:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(linn_map[col]).fillna(df_id_merged[col])
    else:
        df_id_merged[col] = df_id_merged['ID_harmonised'].map(linn_map[col])    

df_id_merged.groupby(['Dataset', 'Stage']).size().unstack()

# Schlesinger

zenodo[zenodo.obs.Project == 'GSM4293555'].obs

# no metadata

df_id_merged.groupby('Dataset').apply(lambda x: x.notna().sum())

df_id_merged['Treatment'].unique()

df_id_merged.CDKN2A_type = df_id_merged.CDKN2A_type.replace('WT', 'Wild_Type')
df_id_merged.SMAD4_type = df_id_merged.SMAD4_type.replace('WT', 'Wild_Type')

# Define the dictionary for harmonizing the 'Location' values
location_dict = {
    'head': 'Head',
    'Head': 'Head',
    'head, uncinate_process': 'Head, Uncinated Process',
    'Head, Uncinated process': 'Head, Uncinated Process',
    'Uncinated process': 'Uncinated Process',
    'neck': 'Neck',
    'neck/body margin': 'Neck/Body',
    'body': 'Body',
    'Body': 'Body',
    'body/tail': 'Body/Tail',
    'Tail': 'Tail',
    'Tail (proximal remanant)': 'Tail, Proximal Remanant',
    'Tail (proximal remanant)': 'Tail, Proximal Remanant',
    'Head/Uncinate': 'Head, Uncinated Process'
}

df_id_merged['Location'] = df_id_merged['Location'].map(location_dict)

df_id_merged['Location'].unique()

# Define the dictionary for harmonizing the 'Grade' values
grade_dict = {
    3.0: 'G3',
    1.0: 'G1',
    2.0: 'G2',
    4: 'G4',
    'G3': 'G3',
    'G2': 'G2',
    'g3': 'G3',
    'g2': 'G2',
    'g2-3': 'G2/G3',
    'gX': 'GX'
}

# Harmonize the 'Grade' column in df_id_merged using the dictionary
df_id_merged['Harmonized_Grade'] = df_id_merged['Grade'].map(grade_dict)

tnm_dict = {
    'ypT2 pN1': 'ypT2 pN1',
    'pT2 pN1': 'pT2 pN1',
    'pT2 pN2': 'pT2 pN2',
    'ypT2 pN0': 'ypT2 pN0',
    'ypT2 pN2': 'ypT2 pN2',
    'pT3 pN2': 'pT3 pN2',
    'ypT3 pN2': 'ypT3 pN2',
    'yT2 pN1': 'yT2 pN1',
    'T4N2M0': 'T4 pN2 M0',
    'T1cN1M0': 'T1c pN1 M0',
    'T2N0M0': 'T2 pN0 M0',
    'T3N0M0': 'T3 pN0 M0',
    'T3N1M0': 'T3 pN1 M0',
    'T1cN2M0': 'T1c pN2 M0',
    'T2N1M0': 'T2 pN1 M0',
    'T3N2M0': 'T3 pN2 M0',
    'T1cN0M0': 'T1c pN0 M0',
    'T2; N1; M0': 'T2 pN1 M0',
    'pT1c, pN2': 'pT1c pN2',
    'T2; N2': 'T2 pN2',
    'T2; N2; M0': 'T2 pN2 M0',
    'T2; N1': 'T2 pN1',
    'ypT2N1M0': 'ypT2 pN1 M0',
    'ypT3N0M0': 'ypT3 pN0 M0',
    'ypT1cN0M0': 'ypT1c pN0 M0',
    'ypT2N0M0': 'ypT2 pN0 M0',
    'ypT1aN1M0': 'ypT1a pN1 M0',
    'ypT2N2M0': 'ypT2 pN2 M0',
    'T2N2M0': 'T2 pN2 M0',
    'ypT1aN0M0': 'ypT1a pN0 M0',
    'ypT1cN1M0': 'ypT1c pN1 M0',
    'ypT3N1M0': 'ypT3 pN1 M0',
    'ypT3 pN0': 'ypT3 pN0',
    'rpT2 moderately differentiated pN1': 'rpT2 pN1',
    'pT2 pN0': 'pT2 pN0',
    'pT1a pN0': 'pT1a pN0',
    'ypT1 moderately differentiated positive for perineural invasion negative for lymphovascular pN0': 'ypT1 pN0'
}

# Harmonize the 'TNM_classification' column using the dictionary
df_id_merged['Harmonized_TNM_Classification'] = df_id_merged['TNM_classification'].map(tnm_dict)

df_id_merged.to_csv('metadata_publications/metadata_all.csv')

# Reload

import pandas as pd
import numpy as np
pd.set_option('display.max_columns', None)

import scanpy as sc

pwd

df_id_merged = pd.read_csv('metadata_publications/metadata_all.csv', index_col='Unnamed: 0')

# change 

df_id_merged.head(2)

df = df_id_merged.groupby('Dataset').apply(lambda x: x.notna().sum())

# Read Anndata

adata = sc.read_h5ad('../Binned_Data/adata_scanvi_neighbours_leiden.h5ad')

## check shared ID_harmonised

shared_ids = adata.obs.groupby('ID_harmonised')['Dataset'].nunique()
shared_ids_multiple_datasets = shared_ids[shared_ids > 1]
shared_ids_multiple_datasets_info = adata.obs[adata.obs['ID_harmonised'].isin(shared_ids_multiple_datasets.index)]

shared_ids_multiple_datasets

shared_ids_multiple_datasets_info.Dataset.unique()

## Resolve by adding prefixes

adata.obs['ID_harmonised'] = adata.obs['ID_harmonised'].astype(str)

conditions = [
    (adata.obs['Dataset'].str.contains('Peng')),
    (adata.obs['Dataset'] == 'Lin')
]
values = ['Peng' + '_' + adata.obs['ID_harmonised'], 'Lin' + '_' + adata.obs['ID_harmonised']]
adata.obs['ID_harmonised'] = np.select(conditions, values, default=adata.obs['ID_harmonised'])

len(set(df_id_merged.ID_harmonised) & set(adata.obs.ID_harmonised))

## recheck

shared_ids = adata.obs.groupby('ID_harmonised')['Dataset'].nunique()
shared_ids_multiple_datasets = shared_ids[shared_ids > 1]
shared_ids_multiple_datasets_info = adata.obs[adata.obs['ID_harmonised'].isin(shared_ids_multiple_datasets.index)]

shared_ids_multiple_datasets

shared_ids_multiple_datasets

# Check if you can add this biological metadata to obsm

adata_obs_subset = adata.obs[['Barcode','Dataset','ID_batch_covariate', 'ID_harmonised', 'Unique_ID', 'Level_1_refined','Tissue', 'Age', 'Sex', 'Diabetes', 'Treatment']].copy()

## Add Steele Treatment

adata_obs_subset.Treatment = np.where(adata_obs_subset.Dataset == 'Steele', 'Treatment Naive', adata_obs_subset.Treatment) 

df_id_merged.drop('Treatment', axis=1, inplace=True)

## can drop Case, Patient_ID, Diabetes

df_id_merged.drop(['Case', 'Patient_ID', 'Diabetes'], axis=1, inplace=True)

## Add Steele and Peng_Normal Age 

adata_obs_subset.Age = np.where(adata_obs_subset.Dataset == 'Steele', '60s', adata_obs_subset.Age) 

adata_obs_subset.Age = np.where(adata_obs_subset.Dataset == 'Steele_Adj_Norm', '60s', adata_obs_subset.Age) 

adata_obs_subset.Age = adata_obs_subset.ID_harmonised.map(dict(zip(df_id_merged[df_id_merged.Dataset == 'Peng'].ID_harmonised, 
         df_id_merged.loc[df_id_merged.Dataset == 'Peng', 'Age']
         .apply(lambda x: f"{int((x // 10) * 10)}s")
         .tolist()))).fillna(adata_obs_subset.Age)

adata_obs_subset.Age = adata_obs_subset.ID_harmonised.map(dict(zip(df_id_merged[df_id_merged.Dataset == 'Peng_Normal'].ID_harmonised, 
         df_id_merged.loc[df_id_merged.Dataset == 'Peng_Normal', 'Age']
         .apply(lambda x: f"{int((x // 10) * 10)}s")
         .tolist()))).fillna(adata_obs_subset.Age)

adata_obs_subset[adata_obs_subset.Dataset.str.contains('Peng')].groupby(['Dataset', 'Age']).size().unstack()

# Drop Age from df_id_merged
df_id_merged.drop('Age', axis=1, inplace=True)

# Drop Tissue
df_id_merged.drop('Tissue', axis=1, inplace=True)

# Drop Dataset, Unique_ID, 
df_id_merged.drop(['Dataset', 'Unique_ID'], axis=1, inplace=True)

biological_metadata = adata_obs_subset.merge(df_id_merged, on='ID_harmonised', how='inner').copy()

import re

[i for i in biological_metadata.columns if re.search(r'_x|_y', i)]

biological_metadata.index = adata.obs.index 

biological_metadata.drop('Barcode', axis=1, inplace=True)

biological_metadata.

df = biological_metadata.groupby('Dataset').nunique() #.apply(lambda x: x.notna().nunique())

counts = {}
for col in df:
    counts[col] = np.count_nonzero(df[col])  
sorted_counts = sorted(counts.items(), key=lambda item: item[1], reverse=True)
for col, count in sorted_counts:
    print(f'{col}: {count}/{len(df[col].index)}')

biological_metadata.groupby(['Dataset', 'Treatment']).size().unstack().T

# Resolved Ding Treatment

biological_metadata[biological_metadata.Dataset.str.contains('Ding')].groupby(['ID_harmonised', 'Treatment']).size().unstack() #.fillna(0).apply(pd.to_numeric, errors='coerce')

ding_ids = biological_metadata[(biological_metadata.Treatment == 'Unknown') & (biological_metadata.Dataset.str.contains('Ding'))].ID_harmonised.unique().tolist() #.groupby('ID_harmonised').size()

len(ding_ids)

ding = pd.read_excel('../Downstream/metadata_publications/Ding_metadata.xlsx', sheet_name='Table1_Clinical_data')

overlapping_ding_id = list(set(ding.Patient_ID) & set(ding_ids))

ding[ding.Patient_ID.isin(overlapping_ding_id)].Patient_ID

dict(zip(ding[ding.Patient_ID.isin(overlapping_ding_id)].Patient_ID, ding[ding.Patient_ID.isin(overlapping_ding_id)].Treatment))

biological_metadata.Treatment.unique().tolist()

ding_treament_dict = {'HT264P1': 'Treatment Naive',
 'HT270P1': 'Treatment Naive',
 'HT284P1': 'Treatment Naive',
 'HT288P1': 'Treatment Naive',
 'HT224P1': 'FOLIRINOX + Gem/Abraxane',
 'HT231P1': 'FOLFIRINOX + SBRT',
 'HT232P1': 'mFOLFIRINOX + FOLFIRI',
 'HT259P1': 'FOLFIRINOX',
 'HT306P1': 'FOLFIRINOX',
 'HT242P1': 'Unknown',
 'HT412P1': 'FOLFIRONO + Gemcitabine'}

biological_metadata.Treatment = biological_metadata.ID_harmonised.map(ding_treament_dict).fillna(biological_metadata.Treatment)

biological_metadata[biological_metadata.Dataset.str.contains('Ding')].groupby(['ID_harmonised', 'Treatment']).size().unstack().fillna(0).apply(pd.to_numeric, errors='coerce').tail(10)

biological_metadata[(biological_metadata.Dataset.str.contains('Ding')) & (biological_metadata.Treatment == 'Unknown')].ID_harmonised.unique()

biological_metadata.Treatment = biological_metadata.Treatment.replace('Untreated', 'Treatment Naive')

biological_metadata.groupby(['Dataset', 'Treatment']).size().unstack().T

# Tissue

biological_metadata.groupby(['Dataset', 'Tissue']).size().unstack().T

biological_metadata.Tissue = np.where(biological_metadata.Dataset.str.contains('Ding'), 'Pancreas', biological_metadata.Tissue)

biological_metadata.Tissue = np.where(biological_metadata.Dataset.str.contains('Peng'), 'Pancreas', biological_metadata.Tissue)

biological_metadata.Tissue = np.where(biological_metadata.Dataset.str.contains('Schlesinger'), 'Pancreas', biological_metadata.Tissue)

biological_metadata.Tissue = np.where(biological_metadata.Dataset.str.contains('Steele'), 'Pancreas', biological_metadata.Tissue)

biological_metadata.groupby(['Dataset', 'Tissue']).size().unstack().T

# Age

biological_metadata.groupby(['Dataset', 'Age']).size().unstack().T

ding_age_ids = list(biological_metadata[(biological_metadata.Dataset.str.contains('Ding')) & (biological_metadata.Age == 'Unknown')].ID_harmonised.unique())

ding.head()

ding_dict_age = dict(zip(ding.Patient_ID, ding.Age.apply(lambda x: f"{int((x // 10) * 10)}s").tolist()))

biological_metadata.Age = biological_metadata.ID_harmonised.map(ding_dict_age).fillna(biological_metadata.Age)

biological_metadata.groupby(['Dataset', 'Age']).size().unstack().T

# Sex

ding_dict_sex = dict(zip(ding.Patient_ID, ding.Sex.apply(lambda x: x.capitalize())))

ding_dict_sex

biological_metadata.Sex = biological_metadata.ID_harmonised.map(ding_dict_sex).fillna(biological_metadata.Sex)

biological_metadata.groupby(['Dataset', 'Sex']).size().unstack().T

peng_sex_dict = {'Peng_N1':'Female',
'Peng_N2':'Male',
'Peng_N3':'Male',
'Peng_N4':'Male',
'Peng_N5':'Female',
'Peng_N6':'Female',
'Peng_N7':'Female',
'Peng_N8':'Male',
'Peng_N9':'Male',
'Peng_N10':'Female',
'Peng_N11':'Female'}

biological_metadata.Sex = biological_metadata.ID_harmonised.map(peng_sex_dict).fillna(biological_metadata.Sex)

biological_metadata.groupby(['Dataset', 'Sex']).size().unstack().T

# TNM Classification

biological_metadata.Harmonized_TNM_Classification.unique()

biological_metadata.Harmonized_TNM_Classification = biological_metadata.Harmonized_TNM_Classification.replace('NaN', 'Unknown')

biological_metadata.groupby(['Dataset', 'Harmonized_TNM_Classification']).size().unstack().T

# Grade

biological_metadata.Grade.unique()

biological_metadata.groupby(['Dataset', 'Grade']).size().unstack().T

grade_mapping = {
    '1.0': 'G1',
    '2.0': 'G2',
    '3.0': 'G3',
    '1': 'G1',
    '2': 'G2',
    '3': 'G3',
    '4': 'G3',  # Assuming grade 4 maps to G3
    'G1': 'G1',
    'G2': 'G2',
    'G3': 'G3',
    'g1': 'G1',
    'g2': 'G2',
    'g3': 'G3',
    'g2-3': 'G2-G3',  # Assuming g2-3 means G2 or G3
    'gX': 'GX',  # Handle unknown grades
    np.nan: 'Unknown'  # Handle missing values
}
biological_metadata['Harmonized_Grade'] = biological_metadata['Grade'].map(grade_mapping)
biological_metadata.groupby(['Dataset', 'Harmonized_Grade']).size().unstack().T

# Stage

biological_metadata.Stage.unique()

biological_metadata.groupby(['Dataset', 'Stage']).size().unstack().T.fillna(0).apply(pd.to_numeric, errors='coerce')

biological_metadata.Stage = biological_metadata.Stage.replace(np.nan, 'Unknown')

stage_mapping = {
    'Ia': 'IA',
    'IB': 'IB',
    'IIA': 'IIA',
    'IIB': 'IIB',
    'IIb': 'IIB',
    'III': 'III',
    np.nan: 'Unknown',  # Handle missing values
}
biological_metadata['Harmonized_Stage'] = biological_metadata['Stage'].map(stage_mapping).fillna('Unknown')
biological_metadata.groupby(['Dataset', 'Harmonized_Stage']).size().unstack().T

simeone_stage = {'P01': 'IV',
'P02': 'IV',
'P03': 'IV',
'P04': 'IB',
'P05': 'III',
'P06': 'III',
'P07': 'IB',
'P08': 'IIB',
'P09': 'IV',
'P10': 'IIB',
'P11': 'IV',
'P12': 'IV',
'P13': 'III',
'P14': 'IB',
'P15': 'III',
'P16': 'IV',
'P17': 'IV',
'P18': 'IV',
'P19': 'IB',
'P20': 'IV',
'P21': 'IV',
'P22': 'III',
'P23': 'IB',
'P24': 'IV',
'P25': 'IV',
'P26': 'IV',
'P26': 'IV',
'P27': 'IV'}

biological_metadata['Harmonized_Stage'] = biological_metadata['ID_harmonised'].map(simeone_stage).fillna(biological_metadata.Harmonized_Stage)
biological_metadata.groupby(['Dataset', 'Harmonized_Stage']).size().unstack().T

df = biological_metadata.groupby('Dataset').nunique() #.apply(lambda x: x.notna().nunique())
counts = {}
for col in df:
    counts[col] = np.count_nonzero(df[col])  
sorted_counts = sorted(counts.items(), key=lambda item: item[1], reverse=True)
for col, count in sorted_counts:
    print(f'{col}: {count}/{len(df[col].index)}')

# Location

biological_metadata.Location.unique()

biological_metadata.groupby(['Dataset', 'Location']).size().unstack().T

peng_location_mapping = {
    'Peng_T1': 'Body',
    'Peng_T2': 'Head',
    'Peng_T3': 'Uncinated Process',
    'Peng_T4': 'Body',
    'Peng_T5': 'Uncinated Process',
    'Peng_T6': 'Tail',
    'Peng_T7': 'Body',
    'Peng_T8': 'Uncinated Process',
    'Peng_T9': 'Head',
    'Peng_T10': 'Uncinated Process',
    'Peng_T11': 'Body/Tail',
    'Peng_T12': 'Uncinated Process',
    'Peng_T13': 'Head',
    'Peng_T14': 'Head',
    'Peng_T15': 'Head',
    'Peng_T16': 'Body',
    'Peng_T17': 'Body/Tail',
    'Peng_T18': 'Body',
    'Peng_T19': 'Head',
    'Peng_T20': 'Head',
    'Peng_T21': 'Head',
    'Peng_T22': 'Body',
    'Peng_T23': 'Head',
    'Peng_T24': 'Head',
    'Peng_N1': 'Tail',
    'Peng_N2': 'Descending Duodenum',
    'Peng_N3': 'Descending Duodenum',
    'Peng_N4': 'Body/Tail',
    'Peng_N5': 'Body/Tail',
    'Peng_N6': 'Body',
    'Peng_N7': 'Tail',
    'Peng_N8': 'Body/Tail',
    'Peng_N9': 'Tail',
    'Peng_N10': 'Common Bile Duct',
    'Peng_N11': 'Body'
}
biological_metadata['Location'] = biological_metadata['ID_harmonised'].map(peng_location_mapping).fillna(biological_metadata.Location)
biological_metadata.groupby(['Dataset', 'Location']).size().unstack().T

biological_metadata.groupby(['Dataset', 'BMI']).size().unstack()

import numpy as np

# Initialize a dictionary to store counts and the datasets for each column
counts = {}
datasets_for_columns = {}

# Loop through columns in the biological_metadata DataFrame
for col in biological_metadata.columns:
    # Exclude 'Unknown' values and get the datasets where values are not 'Unknown'
    non_unknown_data = biological_metadata[biological_metadata[col] != 'Unknown']
    
    # Group by 'Dataset' and count unique values excluding 'Unknown'
    unique_values_per_dataset = non_unknown_data.groupby('Dataset')[col].nunique()

    # Store the number of non-'Unknown' values in each dataset
    counts[col] = np.count_nonzero(unique_values_per_dataset)
    
    # Get the datasets where the value is not 'Unknown'
    datasets_for_columns[col] = unique_values_per_dataset.index.tolist()

# Sort counts in descending order
sorted_counts = sorted(counts.items(), key=lambda item: item[1], reverse=True)

# Print the results
for col, count in sorted_counts:
    print(f'{col}: {count}/{len(biological_metadata["Dataset"].unique())} datasets')
    print(f'Datasets: {", ".join(datasets_for_columns[col])}')

adata.obsm['biological_metadata'] = biological_metadata.copy()

adata.obsm['biological_metadata'] = adata.obsm['biological_metadata'].astype(str)

adata.write('../Binned_Data/adata_scanvi_neighbours_leiden_biological_metadata.h5ad')

adata.obs.Dataset.unique().tolist()

biological_metadata.columns

from collections import Counter

biological_metadata_only_na = biological_metadata.replace('Unknown', np.nan)

df = biological_metadata_only_na.groupby('Dataset').nunique() #.apply(lambda x: x.notna().nunique())
counts = {}
for col in df:
    counts[col] = df[df[col] > 0].index.tolist() #np.count_nonzero(df[col])  

# ANSI escape code for green text
GREEN = '\033[1;32m'
RESET = '\033[0m'

sorted_counts = sorted(counts.items(), key=lambda item: len(item), reverse=False)

for col, count in sorted_counts:
    print(f'{GREEN}{col}{RESET}: {len(count)}/{len(df[col].index)}')
    print(f'Datasets present: {count}')


biological_metadata

ding.groupby('Patient_ID')['Treatment'].unique()

biological_metadata[biological_metadata.Dataset.str.contains('Ding')].groupby('Unique_ID')['Treatment'].unique()

adata
