==> code/human/notebooks/21c_validation_results.ipynb <==
# libraries and data

import warnings
warnings.filterwarnings("ignore")

import scanpy as sc
import pandas as pd
import numpy as np
#from pyclustree import clustree
import matplotlib.pyplot as plt
import os

dfs = []
for file in os.listdir("../../../supplementary_data/human/validations/"):
    if not (file.endswith('.csv') or file.endswith('.csv.gz')):
        continue
    df = pd.read_csv(f"../../../supplementary_data/human/validations/{file}", index_col=0)
    df.index = df.index.astype(int).astype(str)
    dfs.append(df.fillna(""))

import json
with open("../../../supplementary_data/human/mask_mapping.json", "r") as f:
    mask_mapping = json.load(f)

def count_expert_matches(dfs, mask_mapping):
    results = []

    for cluster, true_celltype in mask_mapping.items():
        count = 0

        for df in dfs:


            values = df.loc[str(cluster)]          
            if true_celltype not in values.index:
                continue
            
            if values[true_celltype].lower() == "x":
                count += 1

        results.append({
            "cluster": cluster,
            "true_celltype": true_celltype,
            "num_correct": count
        })

    return pd.DataFrame(results)


results_df = count_expert_matches(dfs, mask_mapping)

# Add groundtruth counts to results_df
results_df['num_correct'] = results_df['num_correct'] + 1
results_df.head()
